cleanup old subject group tables and rename virtual subjects
