+++ app/controllers/api/v1/subjects_controller.rb
@@ -118,21 +118,21 @@ class Api::V1::SubjectsController < Api::ApiController
 
     subject_id_groups = group_selection_result.subject_id_groups
 
-    virtual_subjects = subject_id_groups.each_with_index.map do |ids, idx|
+    subject_groups = subject_id_groups.each_with_index.map do |ids, idx|
       members = Subject.active.where(id: ids).order(Arel.sql("idx(array[#{ids.join(',')}], id)"))
       # Using negative integers as id to avoid clashing with real Subject ids
-      SubjectGroups::VirtualGroupSubject.from_member_subjects(members, virtual_id: -(idx + 1))
+      SubjectGroup.from_member_subjects(members, virtual_id: -(idx + 1))
     end
 
     selection_context = Subjects::SelectorContext.new(
       group_selection_result.subject_selector,
-      virtual_subjects.map(&:id)
+      subject_groups.map(&:id)
     ).format
 
     # Serialize the virtual subjects
-    render json_api: SubjectGroups::VirtualGroupSubjectSelectorSerializer.page(
+    render json_api: SubjectGroupSelectorSerializer.page(
       group_selection_result.subject_selector.params,
-      virtual_subjects,
+      subject_groups,
       selection_context
     )
   end
